OCR/requirements.txt
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
torch==1.1.0
torchvision==0.2.2